Remove fail call from agent on valaidation failure
