+++ 400-rest/src/main/java/software/wings/service/impl/DelegateTaskServiceClassicImpl.java
@@ -839,51 +839,6 @@ public class DelegateTaskServiceClassicImpl implements DelegateTaskServiceClassi
     }
   }
 
-  @Override
-  public void failIfAllDelegatesFailed(
-      final String accountId, final String delegateId, final String taskId, final boolean areClientToolsInstalled) {
-    DelegateTask delegateTask = getUnassignedDelegateTask(accountId, taskId, delegateId);
-    if (delegateTask == null) {
-      log.info("Task {} not found or was already assigned", taskId);
-      return;
-    }
-
-    if (delegateTask.isForceExecute()) {
-      log.debug("Task is set for force execution");
-      return;
-    }
-
-    try (AutoLogContext ignore = new TaskLogContext(taskId, delegateTask.getData().getTaskType(),
-             TaskType.valueOf(delegateTask.getData().getTaskType()).getTaskGroup().name(), OVERRIDE_ERROR)) {
-      if (!isValidationComplete(delegateTask)) {
-        log.info(
-            "Task {} is still being validated with delegate ids {} ", taskId, delegateTask.getValidatingDelegateIds());
-        return;
-      }
-      // Check whether a whitelisted delegate is connected
-      List<String> whitelistedDelegates = assignDelegateService.connectedWhitelistedDelegates(delegateTask);
-      if (isNotEmpty(whitelistedDelegates)) {
-        log.info("Waiting for task {} to be acquired by a whitelisted delegate: {}", taskId, whitelistedDelegates);
-        return;
-      }
-
-      String errorMessage = generateValidationError(delegateTask, areClientToolsInstalled);
-      log.info(errorMessage);
-      DelegateResponseData response;
-      if (delegateTask.getData().isAsync()) {
-        response = ErrorNotifyResponseData.builder()
-                       .failureTypes(EnumSet.of(FailureType.DELEGATE_PROVISIONING))
-                       .errorMessage(errorMessage)
-                       .build();
-      } else {
-        response =
-            RemoteMethodReturnValueData.builder().exception(new InvalidRequestException(errorMessage, USER)).build();
-      }
-      delegateTaskService.processDelegateResponse(accountId, null, taskId,
-          DelegateTaskResponse.builder().accountId(accountId).response(response).responseCode(ResponseCode.OK).build());
-    }
-  }
-
   private String generateValidationError(final DelegateTask delegateTask, final boolean areClientToolsInstalled) {
     final String capabilities = generateCapabilitiesMessage(delegateTask);
     final String delegates = generateValidatedDelegatesMessage(delegateTask);
